refactor: gros refactoring de la partie firebase pour isoler firebase
