+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "assistant-caviste",
-    "version": "5.23.0",
+    "version": "5.24.0",
     "author": "Loic Salou",
     "homepage": "http://ionicframework.com/",
     "private": true,
@@ -76,6 +76,7 @@
         "ionic-plugin-keyboard": "2.2.1",
         "ionicons": "3.0.0",
         "lodash": "4.17.5",
+        "loglevel": "1.6.1",
         "ng2-charts": "1.6.0",
         "phonegap-plugin-barcodescanner": "7.1.0",
         "rxjs": "5.5.7",
