+++ src/_features/admin/login/email-login.page.ts
@@ -1,5 +1,5 @@
 import {Component} from '@angular/core';
-import {LoginService} from '../../../service/login.service';
+import {LoginService} from '../../../service/login/login.service';
 import {NotificationService} from '../../../service/notification.service';
 
 @Component({
