+++ src/_features/racks/cellar-feature.module.ts
@@ -2,14 +2,14 @@ import {NgModule} from '@angular/core';
 import {IonicModule, IonicPageModule} from 'ionic-angular';
 import {CellarPage} from './cellar/cellar.page';
 import {CellarPersistenceService} from '../../service/cellar-persistence.service';
-import {CommonModule} from '@angular/common';
 import {BrowserModule} from '@angular/platform-browser';
 import {SharedModule} from '../../components/shared.module';
 import {FormsModule} from '@angular/forms';
-import {FirebaseConnectionService} from '../../service/firebase-connection.service';
+import {FirebaseAdminService} from '../../service/firebase/firebase-admin.service';
 import {TranslateModule} from '@ngx-translate/core';
 import {UpdateLockerPage} from './update-locker/update-locker.page';
 import {CreateLockerPage} from './create-locker/create-locker.page';
+import {FirebaseLockersService} from '../../service/firebase/firebase-lockers.service';
 
 @NgModule({
             imports: [
@@ -35,8 +35,7 @@ import {CreateLockerPage} from './create-locker/create-locker.page';
               UpdateLockerPage,
             ],
             providers: [
-              CellarPersistenceService,
-              FirebaseConnectionService
+              CellarPersistenceService
             ]
           })
 export class CellarFeatureModule {
