+++ src/model/file-item.ts
