+++ src/model/search-criteria.ts
@@ -1,0 +1,4 @@
+export interface SearchCriteria {
+  keywords: string[];
+  count: number;
+}
