+++ src/service/abstract-persistence.service.ts
@@ -1,5 +1,5 @@
 import {Observable} from 'rxjs/Observable';
-import {LoginService} from './login.service';
+import {LoginService} from './login/login.service';
 import {NotificationService} from './notification.service';
 import {User} from '../model/user';
 import {Subscription} from 'rxjs/Subscription';
@@ -9,11 +9,9 @@ import {TranslateService} from '@ngx-translate/core';
  * Created by loicsalou on 16.06.17.
  */
 
-export abstract class PersistenceService {
+export abstract class AbstractPersistenceService {
 
-  public XREF_ROOT: string;
   protected USERS_FOLDER = 'users';
-  protected XREF_FOLDER = 'xref';
   private loginSub: Subscription;
 
   constructor(protected notificationService: NotificationService, protected loginService: LoginService,
@@ -24,11 +22,9 @@ export abstract class PersistenceService {
   }
 
   protected initialize(user: User) {
-    this.XREF_ROOT = this.XREF_FOLDER;
   }
 
   protected cleanup() {
-    this.XREF_ROOT = undefined;
   }
 
   protected handleError(message: string, error: any) {
