+++ src/service/cave-error.handler.ts
@@ -1,11 +1,11 @@
 import {ErrorHandler, Injectable} from '@angular/core';
-import {FirebaseConnectionService} from './firebase-connection.service';
+import {FirebaseAdminService} from './firebase/firebase-admin.service';
 
 @Injectable()
 export class CaveErrorHandler extends ErrorHandler {
   private count: number = 0;
 
-  constructor(private dataConnection: FirebaseConnectionService) {
+  constructor(private dataConnection: FirebaseAdminService) {
     super();
   }
