+++ src/service/cellar-persistence.service.ts
@@ -6,16 +6,15 @@ import {SimpleLocker} from '../model/simple-locker';
 import {Observable} from 'rxjs';
 import {BehaviorSubject} from 'rxjs/BehaviorSubject';
 import {LockerFactory} from '../model/locker.factory';
-import {LoginService} from './login.service';
-import {PersistenceService} from './persistence.service';
+import {LoginService} from './login/login.service';
+import {AbstractPersistenceService} from './abstract-persistence.service';
 import {NotificationService} from './notification.service';
-import {CellarService} from './cellar.service';
 import {Locker} from '../model/locker';
-import {FirebaseConnectionService} from './firebase-connection.service';
 import {Bottle} from '../model/bottle';
 import {BottlePersistenceService} from './bottle-persistence.service';
 import {TranslateService} from '@ngx-translate/core';
 import {Subscription} from 'rxjs/Subscription';
+import {FirebaseLockersService} from './firebase/firebase-lockers.service';
 
 /**
  * Services related to the cellar itself, locker and place of the lockers.
@@ -23,21 +22,18 @@ import {Subscription} from 'rxjs/Subscription';
  * clicks on a region to filter lockers. Any change on either side must be propagated on the other side.
  */
 @Injectable()
-export class CellarPersistenceService extends PersistenceService implements CellarService {
+export class CellarPersistenceService extends AbstractPersistenceService {
 
   private _lockers: BehaviorSubject<Locker[]> = new BehaviorSubject<Locker[]>([]);
   private _allLockersObservable: Observable<Locker[]> = this._lockers.asObservable();
 
-  get allLockersObservable(): Observable<Locker[]> {
-    return this._allLockersObservable;
-  }
-
   private allLockersArray: Locker[];
   private lockersSubscription: Subscription;
 
-  constructor(private dataConnection: FirebaseConnectionService,
+  constructor(private firebaseLockersService: FirebaseLockersService,
               private lockerFactory: LockerFactory,
               notificationService: NotificationService,
+    // TODO supprimer bottleService quand NGRX fournira le bon sélecteur
               private bottleService: BottlePersistenceService,
               translateService: TranslateService,
               loginService: LoginService) {
@@ -47,11 +43,15 @@ export class CellarPersistenceService extends PersistenceService implements Cell
     }
   }
 
+  get allLockersObservable(): Observable<Locker[]> {
+    return this._allLockersObservable;
+  }
+
   createLocker(locker: Locker): void {
     locker[ 'lastUpdated' ] = new Date().getTime();
-    this.sanitize(locker);
+    sanitize(locker);
     try {
-      this.dataConnection.createLocker(locker);
+      this.firebaseLockersService.createLocker(locker);
     } catch (err) {
       this.notificationService.error('La création du casier a échoué', err)
     }
@@ -59,12 +59,12 @@ export class CellarPersistenceService extends PersistenceService implements Cell
 
   public replaceLocker(locker: SimpleLocker) {
     locker[ 'lastUpdated' ] = new Date().getTime();
-    this.dataConnection.replaceLocker(locker);
+    this.firebaseLockersService.replaceLocker(locker);
   }
 
   public deleteLocker(locker: Locker) {
     if (this.isEmpty(locker)) {
-      this.dataConnection.deleteLocker(locker);
+      this.firebaseLockersService.deleteLocker(locker);
       this.notificationService.information('Le casier "' + locker.name + '" a bien été supprimé')
     } else {
       this.notificationService.warning('Le casier "' + locker.name + '" n\'est pas vide et ne peut donc pas' +
@@ -72,6 +72,7 @@ export class CellarPersistenceService extends PersistenceService implements Cell
     }
   }
 
+  // TODO remplacer pas un selecteur NGRX
   public isEmpty(locker: Locker) {
     return this.bottleService.getBottlesInLocker(locker).length === 0;
   }
@@ -88,33 +89,23 @@ export class CellarPersistenceService extends PersistenceService implements Cell
   }
 
   private initLockers() {
-    this.lockersSubscription = this.dataConnection.fetchAllLockers().subscribe(
+    this.lockersSubscription = this.firebaseLockersService.fetchAllLockers().subscribe(
       (lockers: Locker[]) => {
         this.allLockersArray = lockers.map((locker: Locker) => this.lockerFactory.create(locker));
         this._lockers.next(this.allLockersArray);
       }
     );
   }
-
-  private sanitize(locker: Locker) {
-    locker.dimension = {x: +locker.dimension.x, y: +locker.dimension.y};
-    if (locker[ 'dimensions' ]) {
-      locker[ 'dimensions' ] = locker[ 'dimensions' ].map(
-        dim => {
-          return {x: +dim.x, y: +dim.y}
-        }
-      )
-    }
-  }
-
-  private resolveBottles(bottleIds: string[]): Observable<Bottle[]> {
-    let bottles: Bottle[] = bottleIds.map(id => this.resolveBottle(id)).filter(btl => btl !== undefined);
-    return Observable.create(observer => observer.next(bottles));
-  }
-
-  private resolveBottle(id: string): Bottle {
-    return this.bottleService.getBottle(id);
-  }
 }
 
-
+export function sanitize(locker: Locker): Locker {
+  locker.dimension = {x: +locker.dimension.x, y: +locker.dimension.y};
+  if (locker[ 'dimensions' ]) {
+    locker[ 'dimensions' ] = locker[ 'dimensions' ].map(
+      dim => {
+        return {x: +dim.x, y: +dim.y}
+      }
+    )
+  }
+  return locker;
+}
