+++ src/service/firebase/index.ts
@@ -1,0 +1,6 @@
+export * from './firebase-schema';
+export * from './firebase-admin.service';
+export * from './firebase-bottles.service';
+export * from './firebase-images.service';
+export * from './firebase-lockers.service';
+export * from './firebase-withdrawals.service';
