+++ src/service/login/abstract-login.service.ts
@@ -1,6 +1,6 @@
 import {Observable} from 'rxjs/Observable';
-import {User} from '../model/user';
-import {NotificationService} from './notification.service';
+import {User} from '../../model/user';
+import {NotificationService} from '../notification.service';
 import {Subject} from 'rxjs/Subject';
 import * as firebase from 'firebase/app';
