+++ src/service/login/anonymous-login.service.ts
@@ -4,9 +4,9 @@
 import {Injectable} from '@angular/core';
 import {AngularFireAuth} from 'angularfire2/auth';
 import {AbstractLoginService} from './abstract-login.service';
-import {User} from '../model/user';
+import {User} from '../../model/user';
 import {Observable} from 'rxjs/Observable';
-import {NotificationService} from './notification.service';
+import {NotificationService} from '../notification.service';
 
 /**
  * Services related to the bottles in the cellar.
