+++ src/service/login/login.service.ts
@@ -1,13 +1,13 @@
 import {Observable} from 'rxjs/Observable';
 import {Subject} from 'rxjs/Subject';
-import {User} from '../model/user';
+import {User} from '../../model/user';
 import {AnonymousLoginService} from './anonymous-login.service';
 import {EmailLoginService} from './email-login.service';
 import {FacebookLoginService} from './facebook-login.service';
-import {NotificationService} from './notification.service';
+import {NotificationService} from '../notification.service';
 import {Subscription} from 'rxjs/Subscription';
 import {LocalLoginService} from './local-login.service';
-import {NativeStorageService} from './native-storage.service';
+import {NativeStorageService} from '../native-storage.service';
 import {AbstractLoginService} from './abstract-login.service';
 import {GoogleLoginService} from './google-login.service';
